Clean up AI chat auth and wrapper types
