+++ libs/ai-assistant/core/src/main/kotlin/com/woocommerce/android/aiassistant/core/auth/JwtTokenProvider.kt
@@ -13,9 +13,3 @@ interface JwtTokenProvider {
     suspend fun provide(): String
     suspend fun invalidate() {}
 }
-
-/** Raised by chat auth providers when a transport credential cannot be obtained. */
-class AssistantAuthException(
-    message: String = "Failed to obtain assistant auth credential",
-    cause: Throwable? = null,
-) : RuntimeException(message, cause)
