--- libs/ai-assistant/feature/src/main/kotlin/com/woocommerce/android/aiassistant/chat/woomobileai/WrapperErrorEnvelope.kt
@@ -1,15 +1,0 @@
-package com.woocommerce.android.aiassistant.chat.woomobileai
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-internal data class WrapperErrorEnvelope(
-    val code: String? = null,
-    val message: String? = null,
-    val data: WrapperErrorData? = null,
-)
-
-@Serializable
-internal data class WrapperErrorData(
-    val status: Int? = null,
-)
